move SchemaCache
