+++ src/service/logs/multi.rs
@@ -38,9 +38,8 @@ use crate::{
         distinct_values, get_formatted_stream_name,
         ingestion::{check_ingestion_allowed, evaluate_trigger, write_file, TriggerAlertData},
         logs::StreamMeta,
-        schema::get_upto_discard_error,
+        schema::{get_upto_discard_error, SchemaCache},
         usage::report_request_usage_stats,
-        SchemaCache,
     },
 };
